python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio==1.60.0",
    "grpcio-tools==1.60.0",
    "grpcio-reflection==1.60.0",
    "joblib==1.3.2",
    "numpy==1.26.0",
    "pandas==2.1.4",
    "scikit-learn==1.4.0",
    "scipy>=1.16.3",
    "setuptools==69.0.3",
    "tqdm==4.66.1",
]
